+++ project/package.json
@@ -5,6 +5,7 @@
   "dependencies": {
     "@reduxjs/toolkit": "^1.8.6",
     "axios": "^0.27.2",
+    "classnames": "^2.3.2",
     "history": "^5.3.0",
     "leaflet": "^1.9.3",
     "react": "^18.2.0",
